Refactor address forms
