+++ packages/checkout-storefront/src/components/AddressForm/index.tsx
@@ -1,0 +1,3 @@
+export * from "./AddressForm";
+export * from "./ManualSaveAddressForm";
+export * from "./AutoSaveAddressForm";
