+++ packages/checkout-storefront/src/hooks/__tests__/useGetParsedApiErrors.test.ts
@@ -1,7 +1,7 @@
 import { ApiErrors, useGetParsedApiErrors } from "@/checkout-storefront/hooks/useErrors";
 import { renderHook } from "@testing-library/react-hooks";
 import { apiErrors } from "@/checkout-storefront/lib/fixtures";
-import { AddressFormData } from "@/checkout-storefront/sections/Addresses/types";
+import { AddressFormData } from "@/checkout-storefront/components/AddressForm/types";
 import { CheckoutFormData } from "@/checkout-storefront/sections/CheckoutForm/types";
 
 type TestFormData = AddressFormData & CheckoutFormData;
