+++ packages/checkout-storefront/src/hooks/index.tsx
@@ -10,3 +10,6 @@ export * from "./useOrder";
 export * from "./usePay";
 export * from "./useSetFormErrors";
 export * from "./useCheckoutUpdateStateTrigger";
+export * from "./useAddressAvailability";
+export * from "./useFormDebouncedSubmit";
+export * from "./useAddressFormUtils";
