+++ packages/checkout-storefront/src/hooks/useAddressFormUtils.ts
@@ -2,13 +2,9 @@ import { CountryCode, useAddressValidationRulesQuery } from "@/checkout-storefro
 import { MessageKey, useFormattedMessages } from "@/checkout-storefront/hooks/useFormattedMessages";
 import { AddressField } from "@/checkout-storefront/lib/globalTypes";
 import { warnAboutMissingTranslation } from "@/checkout-storefront/hooks/useFormattedMessages/utils";
-import { reduce } from "lodash-es";
-import {
-  getRequiredAddressFields,
-  getOrderedAddressFields,
-} from "@/checkout-storefront/sections/Addresses/utils";
-import { Address } from "@/checkout-storefront/sections/Addresses/types";
-import { defaultCountry } from "@/checkout-storefront/sections/Addresses/countries";
+import { getRequiredAddressFields, getOrderedAddressFields } from "@/checkout-storefront/lib/utils";
+import { Address } from "@/checkout-storefront/components/AddressForm/types";
+import { defaultCountry } from "@/checkout-storefront/lib/consts";
 
 export const useAddressFormUtils = (countryCode: CountryCode = defaultCountry.code) => {
   const formatMessage = useFormattedMessages();
