+++ packages/checkout-storefront/src/lib/utils/index.ts
@@ -1,3 +1,4 @@
 export * from "./utils";
 export * from "./validation";
 export * from "./money";
+export * from "./address";
