--- packages/checkout-storefront/src/sections/Addresses/index.tsx
@@ -1,3 +1,0 @@
-export * from "./AddressSkeleton";
-export * from "./ShippingAddressSection";
-export * from "./BillingAddressSection";
