+++ packages/checkout-storefront/src/sections/BillingAddressSection/index.tsx
@@ -1,0 +1,1 @@
+export * from "./BillingAddressSection";
