+++ packages/checkout-storefront/src/sections/GuestAddressSection/index.tsx
@@ -1,0 +1,1 @@
+export * from "./GuestAddressSection";
