+++ packages/checkout-storefront/src/sections/ShippingAddressSection/ShippingAddressSection.tsx
@@ -13,10 +13,13 @@ import { CommonSectionProps } from "@/checkout-storefront/lib/globalTypes";
 import { extractMutationErrors } from "@/checkout-storefront/lib/utils";
 import { useAuthState } from "@saleor/sdk";
 import React, { useCallback } from "react";
-import { GuestAddressSection } from "./GuestAddressSection";
-import { AddressFormData, UserAddressFormData } from "./types";
-import { UserAddressSection } from "./UserAddressSection";
-import { getAddressInputData, getAddressVlidationRulesVariables } from "./utils";
+import { GuestAddressSection } from "../GuestAddressSection/GuestAddressSection";
+import { AddressFormData, UserAddressFormData } from "../../components/AddressForm/types";
+import { UserAddressSection } from "../UserAddressSection/UserAddressSection";
+import {
+  getAddressInputData,
+  getAddressVlidationRulesVariables,
+} from "@/checkout-storefront/lib/utils";
 
 export const ShippingAddressSection: React.FC<CommonSectionProps> = ({ collapsed }) => {
   const formatMessage = useFormattedMessages();
