+++ packages/checkout-storefront/src/sections/ShippingAddressSection/index.tsx
@@ -1,0 +1,1 @@
+export * from "./ShippingAddressSection";
