+++ packages/checkout-storefront/src/sections/UserAddressSection/AddressEditForm.tsx
@@ -1,10 +1,13 @@
 import { useErrors } from "@/checkout-storefront/hooks/useErrors";
 import React from "react";
-import { AddressForm, AddressFormProps } from "./AddressForm";
-import { AddressFormData, UserAddressFormData } from "./types";
-import { useAddressList } from "@/checkout-storefront/sections/Addresses/AddressListProvider";
+import { AddressFormData, UserAddressFormData } from "../../components/AddressForm/types";
+import { useAddressList } from "@/checkout-storefront/sections/UserAddressSection/AddressListProvider";
+import {
+  AddressFormProps,
+  ManualSaveAddressForm,
+} from "@/checkout-storefront/components/AddressForm";
 
-interface AddressEditFormProps extends Pick<AddressFormProps<AddressFormData>, "title"> {
+interface AddressEditFormProps extends Pick<AddressFormProps, "title"> {
   defaultValues: UserAddressFormData;
   onClose: () => void;
 }
@@ -17,8 +20,8 @@ export const AddressEditForm: React.FC<AddressEditFormProps> = ({
   const { setApiErrors, ...errorsRest } = useErrors<UserAddressFormData>();
   const { addressUpdate, addressDelete, updating, deleting } = useAddressList();
 
-  const handleUpdate = async (formData: UserAddressFormData) => {
-    const { hasErrors, errors } = await addressUpdate(formData);
+  const handleUpdate = async (formData: AddressFormData) => {
+    const { hasErrors, errors } = await addressUpdate({ ...formData, id: defaultValues.id });
 
     if (!hasErrors) {
       setApiErrors(errors);
@@ -36,12 +39,10 @@ export const AddressEditForm: React.FC<AddressEditFormProps> = ({
   };
 
   return (
-    <AddressForm
+    <ManualSaveAddressForm
       loading={updating || deleting}
       onSubmit={handleUpdate}
-      onDelete={() => {
-        void handleDelete();
-      }}
+      onDelete={handleDelete}
       defaultValues={defaultValues}
       onCancel={onClose}
       {...errorsRest}
